\name{add_char}
\alias{add_char}
\title{Adds quote padding so terms can be searchable with python}
\usage{
  add_char(my_string)
}
\arguments{
  \item{my_string}{A vector of strings that are padded with
  quotation marks so they can be passed to python Required}
}
\value{
  a vector of strings padded with quotes
}
\description{
  Adds quote padding so terms can be searchable with python
}
\author{
  Edmund Hart \email{edmund.m.hart@gmail.com}
}

