\name{format_py_date}
\alias{format_py_date}
\title{Formats rJava strings of dates into text that can be converted to POSIX date}
\usage{
  format_py_date(date_vec)
}
\arguments{
  \item{date_vec}{A vector of dates retured and stripped of
  characters from rJava}
}
\value{
  a vector of POSIX formatted dates
}
\description{
  Formats rJava strings of dates into text that can be
  converted to POSIX date
}
\author{
  Edmund Hart \email{edmund.m.hart@gmail.com}
}

