\name{strip_char}
\alias{strip_char}
\title{Strips rJava S4 class characters from text vector returned from python}
\usage{
  strip_char(Js4text)
}
\arguments{
  \item{Js4text}{vector of rJava S4 class strings}
}
\value{
  a vector stripped of quotes ond brackets
}
\description{
  Strips rJava S4 class characters from text vector
  returned from python
}
\author{
  Edmund Hart \email{edmund.m.hart@gmail.com}
}

